Fixed build after gcs package interface changes.
See jacobsa/gcloud@b0f897fc45c1d3fc682cc3436a051cdb8dd843e8.

+++ gcsproxy/object_proxy.go
@@ -181,7 +181,7 @@ func (op *ObjectProxy) Stat(ctx context.Context) (sr StatResult, err error) {
 
 		sr.Size = fi.Size()
 	} else {
-		sr.Size = op.src.Size
+		sr.Size = int64(op.src.Size)
 	}
 
 	// Stat the object in GCS.
